feat(usertags): Keep tag.key case when updating tags facets
